Refactored google login
