+++ helpers/authUtils.js
@@ -1,7 +1,6 @@
 import { GoogleSignin } from "@react-native-google-signin/google-signin"
-import { mapSocialUserToProfile } from "../../utils/utility";
 
-const authUtil = {
+const authUtils = {
 
     configGoogleAuth: ({ webClientId }) => {
         GoogleSignin.configure({
@@ -36,4 +35,21 @@ const authUtil = {
 
 }
 
-export default authUtil
+const mapSocialUserToProfile = (user) => {
+    if (!user || typeof user !== 'object') {
+        throw new Error('Invalid user object');
+    }
+
+    const { name, email, photo, contact } = user;
+
+    const capitalizedName = name.split(' ').map(chunk => chunk.charAt(0).toUpperCase() + chunk.slice(1)).join(' ')
+
+    return {
+        name: capitalizedName,
+        email: email,
+        contact: contact,
+        image: photo
+    };
+}
+
+export { authUtils, mapSocialUserToProfile }
