+++ screens/auth/SigninWithGoogle/SigninWithGoogle.js
@@ -3,9 +3,9 @@ import {
 } from '@react-native-google-signin/google-signin';
 import { useState } from 'react';
 import { Alert, StyleSheet } from 'react-native';
-import authUtil from '../authUtil';
+import { authUtils } from '../../../helpers/authUtils';
 
-authUtil.configGoogleAuth({
+authUtils.configGoogleAuth({
     webClientId: "174661613395-ddtfki7tpvohilb1htfb86imt4t0is7e.apps.googleusercontent.com"
 });
 
@@ -16,7 +16,7 @@ function SigninWithGoogle({ onSuccess }) {
     const signIn = async () => {
         try {
             setIsInProgress(true)
-            const userInfo = await authUtil.signinUsingGoogle()
+            const userInfo = await authUtils.signinUsingGoogle()
             onSuccess(userInfo)
 
         } catch (error) {
